import doc (#6)
import doc

Convert SMG doc from word and do a lot of renicements
With this PR we are stopping having german version of the document and PDF builds. This is caused by the fact, that documents that we have right now are of a pretty bad quality and we need to first focus on making single English version of it filling our quality expectations. Once this is done we can come back and both translate doc into german and provide PDFs

Reviewed-by: lhajba